Convert the Grid component to use typescript
